refactor(events): split cohost helpers out of _event_helpers.py
_event_helpers.py hit the 500-line CI limit. Extracts the co-host
invite management functions (_can_manage_cohost_invites,
_pending_cohost_invites_out, _update_co_hosts, _update_invited_users)
into a new _event_cohost_helpers.py module.
